requirements.txt
daphne>=4.0.0,<4.1
uwsgi==2.0.22
psycopg2>=2.9.7,<2.10
Django>=4.2.4,<4.3
django-cors-headers>=4.2.0,<4.3
djangorestframework>=3.14.0,<3.15
channels>=4.0.0,<4.1
redis>=5.0.0,<5.1
channels-redis>=4.1.0,<4.2
drf-spectacular>=0.26.4,<0.27
Pillow>=10.0.0,<10.1.0